requirements.txt
langchain-openai==1.1.12
langgraph==1.1.6
streamlit==1.56.0
python-dotenv==1.2.2
langchain-core==1.2.26
openai==2.30.0
